// e2e/scripts/types/macros.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * E2E テスト共通マクロ定義。
 */
export interface Macros {
  [k: string]: {
    /**
     * マクロの概要。
     */
    description?: string;
    /**
     * ステップテキスト展開内容。${ENV_VAR} / <column> 置換記法を含んでよい。
     */
    script: string;
  };
}